+++ app/src/constants/urls.ts
@@ -14,11 +14,6 @@ export const GOOGLE_ANALYTICS_TOS =
 export const GOOGLE_ANALYTICS_PRIVACY_POLICY =
   "https://policies.google.com/privacy?hl=en-US";
 
-export const SPOTIFY_TOS = "https://www.spotify.com/legal/end-user-agreement/";
-export const SPOTIFY_PRIVACY_POLICY =
-  "https://www.spotify.com/legal/privacy-policy/";
-export const SPOTIFY_CONNECTIONS = "https://www.spotify.com/account/apps/";
-
 export const signIn = (lang: string, redirectTo: string) =>
   makeLocalizedUrl(lang, `/sign-in?redirect_to=${redirectTo}`);