+++ app/src/repository/providers/factory.ts
@@ -1,6 +1,5 @@
 import { Provider } from "@/entities/provider";
 import type { ProviderRepositoryInterface } from "@/usecase/interface/provider";
-import { SpotifyProviderRepository } from "./spotify";
 import { YoutubeProviderRepository } from "./youtube";
 
 export function createProviderRepository(
@@ -9,8 +8,6 @@ export function createProviderRepository(
   switch (type) {
     case Provider.GOOGLE:
       return new YoutubeProviderRepository();
-    case Provider.SPOTIFY:
-      return new SpotifyProviderRepository();
     default:
       throw new Error(`Provider type "${type}" is not implemented.`);
   }