+++ packages/spotify/src/index.ts
@@ -1,6 +1,0 @@
-export { ApiClient, ApiClientOptions } from "./ApiClient";
-export { PlaylistManager } from "./managers/PlaylistManager";
-export { IPage, Page } from "./Page";
-export { SpotifyApiError } from "./SpotifyApiError";
-export { IImage, IImages, Image, Images } from "./structures/Image";
-export { IPlaylist, Playlist, RawPlaylist } from "./structures/Playlist";